Bug fixes for log recovery log messages. They now report total bytes sent as well as the rate at which bytes are sent

+++ src/frontend/org/voltdb/VoltDBInterface.java
@@ -133,5 +133,5 @@ public interface VoltDBInterface
     /**
      * Notify RealVoltDB that recovery is complete
      */
-    void onRecoveryCompletion(double transferred);
+    void onRecoveryCompletion(long transferred);
 }
